Move add item into a drawer component
