+++ app/router.js
@@ -6,11 +6,13 @@ import ProductSelector from './views/pages/product-selector';
 import Items from './views/pages/items';
 import ItemDetail from './views/pages/item-detail';
 import Search from './views/pages/search';
+import AddItem from './views/pages/add-item';
 
 var routes = (
   <Route name="app" path="/" handler={App}>
     <Route name="product" path="product/:id" handler={Items} >
       <Route name="item/:number" handler={ItemDetail} />
+      <Route name="add-item" path="add-item" handler={AddItem} />
     </Route>
     <Route name="search" path="search" handler={Search} />
     <DefaultRoute handler={ProductSelector} />
