+++ app/views/components/item-detail/item-attachments.js
