+++ app/views/components/item-detail/item-description.js
