+++ app/views/components/item-detail/item-details.js
