+++ app/views/components/item-detail/item-title.js
