+++ app/views/pages/item-detail.js
@@ -2,21 +2,21 @@ import React from 'react/addons';
 import _ from 'lodash';
 
 // Stores
-import ProductStore from '../../../stores/product-store';
-import ItemActions from '../../../actions/item-actions';
+import ProductStore from '../../stores/product-store';
+import ItemActions from '../../actions/item-actions';
 // Components
 import ItemHeader from './item-header';
-import ItemDetails from './item-details';
-import ItemDescription from './item-description';
-import ItemAttachments from './item-attachments';
-import ItemSubitems from './item-subitems';
-import ItemComments from './item-comments';
-import ItemActivity from './item-activity';
-import ItemDetailMixin from './detail-mixin';
+import ItemDetails from '../components/item-detail/item-details';
+import ItemDescription from '../components/item-detail/item-description';
+import ItemAttachments from '../components/item-detail/item-attachments';
+import ItemSubitems from '../components/item-detail/item-subitems';
+import ItemComments from '../components/item-detail/item-comments';
+import ItemActivity from '../components/item-detail/item-activity';
+import ItemDetailMixin from '../components/item-detail/detail-mixin';
 // Libs
 import {State,Link} from 'react-router';
 
-let initialItemDetailHeight = function() {
+let stripeHeight = function() {
   let bodyHeight = document.body.getBoundingClientRect().height;
   let headerHeight = document.getElementsByClassName('product__header-menu')[0].getBoundingClientRect().height;
 
@@ -33,18 +33,18 @@ var ItemDetail = React.createClass({
     return {
       item: {},
       product: product,
-      itemDetailHeight: initialItemDetailHeight(),
-      descriptionEditable: false,
-      attachmentsOpen: false
+      attachmentsPanel: false,
+      stripeHeight: stripeHeight(),
+      descriptionEditable: false
     };
   },
 
   updateStripeHeight() {
-    let content = document.getElementsByClassName('content__wrapper')[0];
+    let content = document.getElementsByClassName('drawer__content')[0];
     let height = content ? content.getBoundingClientRect().height : 0;
 
-    if (height > this.state.itemDetailHeight) {
-      this.setState({itemDetailHeight: height})
+    if (height > this.state.stripeHeight) {
+      this.setState({stripeHeight: height})
     }
   },
 
@@ -147,6 +147,20 @@ var ItemDetail = React.createClass({
     )
   },
 
+  drawerStripe() {
+    let stripeClass = `stripe ${this.state.item.type}`;
+    let closeClass = `drawer__close ${this.state.item.type}`;
+    let stripeStyles = {height: `${this.state.stripeHeight}px`};
+
+    return (
+      <div style={stripeStyles} className={stripeClass}>
+        <Link to="product" params={{ id: this.getParams().id }} className={closeClass}>
+          <span aria-hidden="true" className="glyphicon glyphicon-remove"/>
+        </Link>
+      </div>
+    )
+  },
+
   componentWillReceiveProps(nextProps) {
     this.updateStripeHeight();
 
@@ -174,10 +188,6 @@ var ItemDetail = React.createClass({
       return <div/>;
     }
 
-    let stripeClass = `stripe ${this.state.item.type}`;
-    let closeClass = `item-detail__close ${this.state.item.type}`;
-    var stripeStyles = {height: `${this.state.itemDetailHeight}px`};
-
     let itemDetails = this.itemDetails()
     let itemDescription = this.itemDescription();
 
@@ -198,19 +208,15 @@ var ItemDetail = React.createClass({
     let itemActivity = this.itemActivity()
 
     return (
-      <div ref="itemDetail" className="container-fluid item-detail no-gutter">
-        <div style={stripeStyles} className={stripeClass}>
-          <Link to="product" params={{ id: this.getParams().id }} className={closeClass}>
-            <span aria-hidden="true" className="glyphicon glyphicon-remove"/>
-          </Link>
-        </div>
-        <div className="content__wrapper">
-          {itemDetails}
-          {itemDescription}
-          {itemAttachments}
-          {subitems}
-          {itemComments}
-          {itemActivity}
+      <div ref="itemDetail" className="container-fluid item-detail no-gutter drawer">
+        {this.drawerStripe()}
+        <div className="drawer__content">
+          {this.itemDetails()}
+          {this.itemDescription()}
+          {this.itemAttachments()}
+          {this.itemSubitems()}
+          {this.itemComments()}
+          {this.itemActivity()}
         </div>
       </div>
     )
