--- public/less/components/add-item-modal.less
@@ -1,344 +1,0 @@
-.add-item.modal {
-  padding-top: 8em;
-
-  .modal-body {
-    .transition(opacity 200ms);
-    background: @sprintly-background;
-    border-radius: 0px 4px 4px 4px;
-
-    &.dragging {
-      .opacity(0.8);
-      position: relative;
-      &::after {
-        content: 'Drop to attach.';
-        position: absolute;
-        top: 30%;
-        left: 0;
-        width: 100%;
-        text-align: center;
-        font-weight: bold;
-        font-size: 1.4em;
-      }
-    }
-  }
-
-  .modal-content {
-    background: none;
-    border: none;
-    box-shadow: none;
-    overflow: auto!important;
-  }
-}
-
-
-.add-item .btn-primary {
-  background: #84b431;
-  color: #fff;
-  border: none;
-}
-
-.add-item .selector {
-  margin-top: 1.4em;
-  min-height: 36px;
-
-  .selector__wrapper {
-    position: fixed;
-    max-height: 22em;
-    overflow: scroll;
-  }
-}
-
-.add-item__tabs {
-  position: relative;
-  border: none;
-
-  li {
-    a {
-      color: #fff;
-      padding: 5px 25px;
-      font-size: 1.2em;
-      border: none;
-    }
-
-    &.add-item__nav-story > a {
-      background-color: @story-color;
-
-      &:hover {
-        background-color: lighten(@story-color, 12%);
-      }
-    }
-
-    &.add-item__nav-task > a {
-      background-color: @task-color;
-
-      &:hover {
-        background-color: lighten(@task-color, 12%);
-      }
-    }
-
-    &.add-item__nav-defect > a {
-      background-color: @defect-color;
-
-      &:hover {
-        background-color: lighten(@defect-color, 12%);
-      }
-    }
-
-    &.add-item__nav-test > a {
-      background-color: @test-color;
-
-      &:hover {
-        background-color: lighten(@test-color, 12%);
-      }
-    }
-
-    &.active > a,
-    &.active > a:hover {
-      background: @sprintly-background;
-    }
-  }
-}
-
-.item-title__field,
-.add-item__field {
-  > span {
-    color: @sprintly-blue;
-  }
-}
-
-.add-item__field {
-  > span {
-    font-size: 1.4em;
-  }
-}
-
-.item-title__field {
-  > span {
-    font-size: 20px;
-  }
-
-  @media (max-width: @screen-xs-min) {
-    font-size: 14px;
-  }
-}
-
-.item-title__field {
-  margin-bottom: 1.8em;
-}
-.add-item__field {
-  margin-bottom: 2em;
-}
-.item-title__field,
-.add-item__field {
-  display: inline-block;
-
-  .input-group {
-    position: relative;
-    display: inline-block;
-  }
-
-  label {
-    position: absolute;
-    left: 38%;
-    bottom: -36px;
-    font-size: 11px;
-    font-weight: normal;
-  }
-
-  input.form-control {
-    border: none;
-    border-radius: 0;
-    display: inline-block;
-    margin: 0 8px -11px;
-    font-size: 1.2em;
-    border-bottom: 1px dashed #444;
-
-    &.invalid {
-      border-bottom: 2px solid @defect-color;
-    }
-  }
-}
-
-.add-item__field.who input.form-control {
-  width: 220px;
-}
-.add-item__field.what input.form-control {
-  width: 373px;
-}
-.add-item__field.why input.form-control {
-  width: 660px;
-}
-
-.item-title__field.who {
-  width: 35%;
-  @media (max-width: @screen-xs-min) {
-    width: 100%;
-  }
-}
-.item-title__field.what {
-  width: 65%;
-
-  @media (max-width: @screen-xs-min) {
-    width: 100%;
-  }
-}
-.item-title__field.default,
-.item-title__field.why {
-  width: 100%;
-  margin-bottom: 10px;
-}
-.item-title__field.who .input-group {
-  position: relative;
-  display: inline-block;
-  width: 68%;
-
-  @media (max-width: @screen-xs-min) {
-    width: 76%;
-  }
-}
-.item-title__field.what .input-group {
-  position: relative;
-  display: inline-block;
-  width: 82%;
-
-  @media (max-width: @screen-xs-min) {
-    width: 74%;
-  }
-}
-.item-title__field.why .input-group {
-  position: relative;
-  display: inline-block;
-  width: 87%;
-
-  // TODO: UI on mobile will require a different design
-  @media (max-width: @screen-xs-min) {
-    width: 71%;
-  }
-}
-.item-title__field.default .input-group {
-  position: relative;
-  display: inline-block;
-  width: 100%;
-}
-
-.add-item__actions {
-  margin: 1em 0;
-
-  .btn {
-    margin-right: 5px;
-  }
-
-  .cancel-item,
-  .create-item,
-  .checkbox {
-    float: right;
-  }
-
-  .create-item {
-    border: 1px solid @story-color;
-  }
-
-  @media (min-width: 768px) {
-    .create-item {
-      margin-right: 10px;
-    }
-  }
-}
-
-@media (min-width: 768px) {
-  .modal-dialog {
-    width: 768px;
-  }
-}
-
-.add-item__member-dropdown {
-  margin-top: 1em;
-
-  // Override react-select
-  .Select-menu {
-    max-height: 72px;
-  }
-}
-
-.add-item.modal .react-mentions {
-  position: relative;
-  overflow-y: visible;
-
-  .control {
-    &.textarea {
-      font-weight: normal;
-      background: #fff;
-      border: 1px solid #ccc;
-      border-radius: 4px;
-
-      textarea {
-        min-height: 6em;
-        width: 100%;
-        padding: 6px 12px;
-        width: 100%;
-        display: block;
-        margin: 0;
-        border: none;
-        background: transparent;
-        position: relative;
-        outline: 0;
-        overflow: hidden;
-        resize: vertical !important;
-      }
-
-      .highlighter {
-        padding: 6px 12px;
-        width: inherit;
-        position: absolute;
-        color: transparent;
-        overflow: hidden;
-      }
-    }
-
-    .highlighter {
-      width: inherit;
-      position: absolute;
-      color: transparent;
-      overflow: hidden;
-      font-size: 14px;
-
-      strong {
-        font-weight: normal;
-        background: #cee4e5;
-        color: #444;
-      }
-    }
-  }
-
-  .suggestions {
-    position: absolute;
-    z-index: @zIndex-2;
-    max-height: 260px;
-    overflow-y: hidden;
-
-    border: 1px solid rgba(0,0,0,0.15);
-    background-color: white;
-    margin-top: 14px;
-    font-size: 10pt;
-
-    ul {
-      margin: 0;
-      padding: 0;
-      list-style-type: none;
-
-      li {
-        padding: 5px 15px;
-        cursor: pointer;
-        border-bottom: 1px solid rgba(0,0,0,0.15);;
-
-        &:last-child {
-          border-bottom: none;
-        }
-
-        &.focus {
-          background-color: #cee4e5;
-        }
-      }
-    }
-  }
-}
