+++ public/less/components/drawer.less
@@ -1,0 +1,159 @@
+.drawer {
+  position: fixed;
+  height: 100%;
+
+  background: #fff;
+  top: 54px;
+  right: 0;
+  z-index: @zIndex-3;
+  overflow: scroll;
+  box-shadow: 0px 0px 20px 2px rgba(0, 0, 0, 0.25);
+
+  @media (max-width: @screen-xs-min) {
+    width: 100%;
+  }
+  @media (min-width: @screen-sm-min) {
+    width: 80%;
+  }
+  @media (min-width: @screen-md-min) { // < 992px
+    width: 70%;
+  }
+  @media (min-width: @screen-lg-min) and (max-width: @screen-xlg-min) { // 1200px < x > 1600px
+    width: 60%;
+  }
+  @media (min-width: @screen-lg-min) { // < 1200px
+    width: 66%;
+  }
+  @media (min-width: @screen-xlg-min) { // < 1600px
+    width: 44%;
+  }
+
+  -webkit-animation: myanim 300ms ease-out;
+          animation: myanim 300ms ease-out;
+}
+@-webkit-keyframes myanim {
+  from {
+    .translate(100%, 0);
+  }
+  to {
+    .translate(0, 0);
+  }
+}
+@keyframes myanim {
+  from {
+    .translate(100%, 0);
+  }
+  to {
+    .translate(0, 0);
+  }
+}
+.drawer__close {
+  width: 100%;
+  display: inline-block;
+  margin: 0px;
+  text-align: center;
+  line-height: 52px;
+  border-radius: 0px;
+  color: white;
+  font-weight: 600;
+  font-size: 16px;
+  padding-top: 12px;
+  height: 75px;
+
+  @media (max-width: @screen-sm-min) {
+    font-size: 20px;
+  }
+
+  @media (max-width: @screen-xs-min) {
+    font-size: 14px;
+  }
+}
+.drawer__close.story {
+  background-color: @story-color;
+}
+.drawer__close.defect {
+  background-color: @defect-color;
+}
+.drawer__close.task {
+  background-color: @task-color;
+}
+.drawer__close.test {
+  background-color: @test-color;
+}
+.drawer__content {
+  width: 97.5%;
+  display: inline-block;
+
+  @media (max-width: @screen-xs-min) {
+    width: 90%;
+  }
+  @media (min-width: @screen-sm-min) {
+    width: 95.5%;
+  }
+}
+
+.drawer .react-mentions {
+  .textarea {
+    min-height: 6em;
+    border: 1px solid @sprintly-black-light;
+    border-radius: 4px;
+  }
+
+  textarea {
+    position: absolute;
+    top: 0;
+    padding: 6px 12px;
+    outline: 0;
+    border: 0;
+    margin: 0;
+  }
+
+  .highlighter {
+    width: inherit;
+    color: transparent;
+    font: inherit;
+    overflow: hidden;
+    white-space: pre-wrap;
+    word-wrap: break-word;
+    padding: 9px;
+
+    strong {
+      font-weight: normal;
+      background: #cee4e5;
+      padding-top: 5px;
+      padding-right: 5px;
+    }
+  }
+
+  .suggestions {
+    position: absolute;
+    z-index: @zIndex-3;
+    max-height: 260px;
+    overflow-y: hidden;
+
+    border: 1px solid rgba(0,0,0,0.15);
+    background-color: white;
+    margin-top: 14px;
+    font-size: 10pt;
+
+    ul {
+      margin: 0;
+      padding: 0;
+      list-style-type: none;
+
+      li {
+        padding: 5px 15px;
+        cursor: pointer;
+        border-bottom: 1px solid rgba(0,0,0,0.15);;
+
+        &:last-child {
+          border-bottom: none;
+        }
+
+        &.focus {
+          background-color: #cee4e5;
+        }
+      }
+    }
+  }
+}
