+++ public/less/components/item-detail/index.less
@@ -1,58 +1,3 @@
-.item-detail {
-  position: fixed;
-  height: 100%;
-
-  @media (max-width: @screen-xs-min) {
-    width: 100%;
-  }
-  @media (min-width: @screen-sm-min) {
-    width: 80%;
-  }
-  @media (min-width: @screen-md-min) { // < 992px
-    width: 70%;
-  }
-  @media (min-width: @screen-lg-min) and (max-width: @screen-xlg-min) { // 1200px < x > 1600px
-    width: 60%;
-  }
-  @media (min-width: @screen-lg-min) { // < 1200px
-    width: 66%;
-  }
-  @media (min-width: @screen-xlg-min) { // < 1600px
-    width: 44%;
-  }
-
-  background: #fff;
-  top: 54px;
-  right: 0;
-  z-index: 299;
-  overflow: scroll;
-  box-shadow: 0px 0px 20px 2px rgba(0, 0, 0, 0.25);
-
-  .close {
-    position: absolute;
-    top: 1em;
-    right: 1em;
-  }
-
-  -webkit-animation: myanim 300ms ease-out;
-          animation: myanim 300ms ease-out;
-}
-@-webkit-keyframes myanim {
-  from {
-    .translate(100%, 0);
-  }
-  to {
-    .translate(0, 0);
-  }
-}
-@keyframes myanim {
-  from {
-    .translate(100%, 0);
-  }
-  to {
-    .translate(0, 0);
-  }
-}
 .item-detail .item-card {
   margin-bottom: 2em;
 }
@@ -60,77 +5,9 @@
   float: left;
   width: 100%;
 }
-.item-detail .stripe {
-  width: 2.5%;
-  float: left;
-  display: inline-block;
-
-  @media (min-width: @screen-sm-min) {
-    width: 3.5%;
-  }
-  @media (max-width: @screen-xs-min) {
-    width: 10%;
-  }
-}
-.item-detail .stripe.story {
-  background-color: @story-bg;
-}
-.item-detail .stripe.defect {
-  background-color: @defect-bg;
-}
-.item-detail .stripe.task {
-  background-color: @task-bg;
-}
-.item-detail .stripe.test {
-  background-color: @test-bg;
-}
 .item-detail .header {
   position: relative;
   z-index: @zIndex-1;
-}
-.item-detail .content__wrapper {
-  width: 97.5%;
-  display: inline-block;
-
-  @media (max-width: @screen-xs-min) {
-    width: 90%;
-  }
-  @media (min-width: @screen-sm-min) {
-    width: 95.5%;
-  }
-}
-.item-detail .item-detail__close {
-  width: 100%;
-  display: inline-block;
-  margin: 0px;
-  text-align: center;
-  line-height: 52px;
-  border-radius: 0px;
-  color: white;
-  font-weight: 600;
-  font-size: 16px;
-  padding-top: 12px;
-  height: 75px;
-
-  @media (max-width: @screen-sm-min) {
-    font-size: 20px;
-  }
-
-  @media (max-width: @screen-xs-min) {
-    font-size: 14px;
-  }
-}
-.item-detail .item-detail__close.story {
-  background-color: @story-color;
-}
-.item-detail .item-detail__close.defect {
-  background-color: @defect-color;
-}
-.item-detail .item-detail__close.task {
-  background-color: @task-color;
-}
-.item-detail .item-detail__close.test {
-  background-color: @test-color;
 }
 .item-detail .section {
   .no-gutter;
@@ -301,70 +178,5 @@ li.action__restricted {
 
   span.status {
     font-weight: @fontWeight-6;
-  }
-}
-.item-detail .react-mentions {
-  .textarea {
-    min-height: 6em;
-    border: 1px solid @sprintly-black-light;
-    border-radius: 4px;
-  }
-
-  textarea {
-    position: absolute;
-    top: 0;
-    padding: 6px 12px;
-    outline: 0;
-    border: 0;
-    margin: 0;
-  }
-
-  .highlighter {
-    width: inherit;
-    color: transparent;
-    font: inherit;
-    overflow: hidden;
-    white-space: pre-wrap;
-    word-wrap: break-word;
-    padding: 9px;
-
-    strong {
-      font-weight: normal;
-      background: #cee4e5;
-      padding-top: 5px;
-      padding-right: 5px;
-    }
-  }
-
-  .suggestions {
-    position: absolute;
-    z-index: @zIndex-2;
-    max-height: 260px;
-    overflow-y: hidden;
-
-    border: 1px solid rgba(0,0,0,0.15);
-    background-color: white;
-    margin-top: 14px;
-    font-size: 10pt;
-
-    ul {
-      margin: 0;
-      padding: 0;
-      list-style-type: none;
-
-      li {
-        padding: 5px 15px;
-        cursor: pointer;
-        border-bottom: 1px solid rgba(0,0,0,0.15);;
-
-        &:last-child {
-          border-bottom: none;
-        }
-
-        &.focus {
-          background-color: #cee4e5;
-        }
-      }
-    }
   }
 }
