+++ public/less/components/item-detail/item-title.less
@@ -1,0 +1,100 @@
+.item-title__field {
+  > span {
+    color: @sprintly-blue;
+  }
+}
+.item-title__field {
+  > span {
+    font-size: 20px;
+  }
+
+  @media (max-width: @screen-xs-min) {
+    font-size: 14px;
+  }
+}
+
+.item-title__field {
+  margin-bottom: 1.8em;
+}
+
+.item-title__field,
+.add-item__field {
+  display: inline-block;
+
+  .input-group {
+    position: relative;
+    display: inline-block;
+  }
+
+  label {
+    position: absolute;
+    left: 38%;
+    bottom: -36px;
+    font-size: 11px;
+    font-weight: normal;
+  }
+
+  input.form-control {
+    border: none;
+    border-radius: 0;
+    display: inline-block;
+    margin: 0 8px -11px;
+    font-size: 1.2em;
+    border-bottom: 1px dashed #444;
+
+    &.invalid {
+      border-bottom: 2px solid @defect-color;
+    }
+  }
+}
+.item-title__field.who {
+  width: 35%;
+  @media (max-width: @screen-xs-min) {
+    width: 100%;
+  }
+}
+.item-title__field.what {
+  width: 65%;
+
+  @media (max-width: @screen-xs-min) {
+    width: 100%;
+  }
+}
+.item-title__field.default,
+.item-title__field.why {
+  width: 100%;
+  margin-bottom: 10px;
+}
+.item-title__field.who .input-group {
+  position: relative;
+  display: inline-block;
+  width: 68%;
+
+  @media (max-width: @screen-xs-min) {
+    width: 76%;
+  }
+}
+.item-title__field.what .input-group {
+  position: relative;
+  display: inline-block;
+  width: 82%;
+
+  @media (max-width: @screen-xs-min) {
+    width: 74%;
+  }
+}
+.item-title__field.why .input-group {
+  position: relative;
+  display: inline-block;
+  width: 87%;
+
+  // TODO: UI on mobile will require a different design
+  @media (max-width: @screen-xs-min) {
+    width: 71%;
+  }
+}
+.item-title__field.default .input-group {
+  position: relative;
+  display: inline-block;
+  width: 100%;
+}
