+++ public/less/components/tags-input.less
@@ -49,11 +49,10 @@
     left: 0;
     position: absolute;
     top: 33px;
-    width: 100.3%;
+    width: 100%;
     font-size: 11px;
     z-index: 1;
     overflow-y: auto;
-    max-height: 109px;
     margin-top: -1px;
     margin-left: -1px;
     margin-right: -1px;
