+++ public/less/main.less
@@ -12,6 +12,8 @@
 
 // Components
 @import "components/header";
+@import "components/drawer";
+@import "components/stripe";
 @import "components/columns";
 @import "components/buttons";
 @import "components/item-detail/index";
@@ -33,7 +35,7 @@
 @import "components/filters-toolbar";
 @import "components/product-dropdown";
 @import "components/search";
-@import "components/add-item-modal";
+@import "components/add-item";
 @import "components/tags-input";
 @import "components/sidebars/sidebar-filters";
 @import "overrides";
