+++ test/index.js
@@ -5,7 +5,6 @@ require("../app/actions/velocity-actions-test.js");
 require("../app/stores/filters-store-test.js");
 require("../app/stores/product-store-test.js");
 require("../app/stores/sidebar-store-test.js");
-require("../app/views/components/add-item-modal-test.js");
 require("../app/views/components/filters/filter-component-test.js");
 require("../app/views/components/filters/forms/members-filter-test.js");
 require("../app/views/components/filters/velocity-component-test.js");
@@ -17,6 +16,7 @@ require("../app/views/components/item-column/sprint-group-test.js");
 require("../app/views/components/sidebars/filters-test.js");
 require("../app/views/components/sidebars/index-test.js");
 require("../app/views/components/tags-input-test.js");
+require("../app/views/pages/add-item-test.js");
 require("../app/views/pages/helpers-test.js");
 require("../app/views/pages/items-test.js");
 require("../app/views/pages/search-test.js");
